\name{whatsup-package}
\alias{whatsup-package}
\alias{whatsup}
\docType{package}
\title{
Astronomical Observability
}
\description{
Produces observability plots for arbitrary locations, dates and times.
}
\details{
\tabular{ll}{
Package: \tab whatsup\cr
Type: \tab Package\cr
Version: \tab 0.1.3\cr
Date: \tab 2021-11-22\cr
License: \tab LGPL-3\cr
}
}
\author{
Aaron Robotham

Maintainer: Aaron Robotham <aaron.robotham@uwa.edu.au>
}

\keyword{ package }
\keyword{ whatsup }


