\name{whatsup-package}
\alias{whatsup-package}
\alias{whatsup}
\docType{package}
\title{
Astronomical Observability
}
\description{
Produces observability plots for arbitrary locations, dates and times.
}
\details{
\tabular{ll}{
Package: \tab whatsup\cr
Type: \tab Package\cr
Version: \tab 0.1.1\cr
Date: \tab 2019-03-18\cr
License: \tab GPL-3\cr
}
}
\author{
Aaron Robotham

Maintainer: Aaron Robotham <aaron.robotham@uwa.edu.au>
}

\keyword{ package }
\keyword{ whatsup }


